Rework the basics and tie into the testserver
