+++ edx_notifications/channels/triggered_email.py
@@ -4,13 +4,17 @@ that sends email to the users.
 """
 
 import logging
-from edx_notifications.channels.durable import BaseDurableNotificationChannel
+from edx_notifications.channels.channel import BaseNotificationChannelProvider
 from edx_notifications.scopes import resolve_user_scope
+
+from edx_notifications.data import UserNotification
+
+from edx_notifications.channels.link_resolvers import MsgTypeToUrlResolverMixin
 
 log = logging.getLogger(__name__)
 
 
-class TriggeredEmailChannelProvider(BaseDurableNotificationChannel):
+class TriggeredEmailChannelProvider(MsgTypeToUrlResolverMixin, BaseNotificationChannelProvider):
     """
     A TriggeredEmail notification channel will
     send email to the user.
@@ -21,10 +25,28 @@ class TriggeredEmailChannelProvider(BaseDurableNotificationChannel):
         Send a notification to a user, which - in a TriggerEmailChannel Notification
         """
 
-        # user_ids = resolve_user_scope('student_email_resolver', {'user_id': user_id})
-        msg = super(TriggeredEmailChannelProvider, self)._get_linked_resolved_msg(msg)
+        # call into one of the registered resolvers to get the email for this
+        # user
+        scope_results = resolve_user_scope('student_email_resolver', {'user_id': user_id})
 
-        return msg
+        msg = self._get_linked_resolved_msg(msg)
+
+        user_msg = UserNotification(
+            user_id=user_id,
+            msg=msg
+        )
+
+        for result in scope_results:
+            #
+            # Do the rendering and the sending of the email
+            #
+            if isinstance(result, dict):
+                email = result['email']
+            else:
+                email = result
+            print '***** email = {}'.format(email)
+
+        return user_msg
 
     def bulk_dispatch_notification(self, user_ids, msg, exclude_user_ids=None, channel_context=None):
         """
@@ -34,18 +56,10 @@ class TriggeredEmailChannelProvider(BaseDurableNotificationChannel):
         user_ids should be a list, a generator function, or a django.db.models.query.ValuesListQuerySet
         when directly feeding in a Django ORM queryset, where we select just the id column of the user
         """
-        exclude_user_ids = exclude_user_ids if exclude_user_ids else []
 
         # enumerate through the list of user_ids and call
         # dispatch_notification_to_user method.
         #  e sure not to include any user_id in the exclude list
         for user_id in user_ids:
-            if user_id not in exclude_user_ids:
+            if not exclude_user_ids or user_id not in exclude_user_ids:
                 self.dispatch_notification_to_user(user_id, msg, channel_context)
-
-    def resolve_msg_link(self, msg, link_name, params, channel_context=None):
-        """
-        Generates the appropriate link given a msg, a link_name, and params
-        """
-        resolved_link = super(TriggeredEmailChannelProvider, self).resolve_msg_link(msg, link_name, params, channel_context=None)
-        return resolved_link
