+++ testserver/settings.py
@@ -145,6 +145,23 @@ NOTIFICATION_CLICK_LINK_URL_MAPS = {
 # list all known channel providers
 NOTIFICATION_CHANNEL_PROVIDERS = {
     'durable': {
+        'class': 'edx_notifications.channels.durable.BaseDurableNotificationChannel',
+        'options': {
+            # list out all link resolvers
+            'link_resolvers': {
+                # right now the only defined resolver is 'type_to_url', which
+                # attempts to look up the msg type (key) via
+                # matching on the value
+                'msg_type_to_url': {
+                    'class': 'edx_notifications.channels.link_resolvers.MsgTypeToUrlLinkResolver',
+                    'config': {
+                        '_click_link': NOTIFICATION_CLICK_LINK_URL_MAPS,
+                    }
+                }
+            }
+        }
+    },
+    'triggered-email': {
         'class': 'edx_notifications.channels.triggered_email.TriggeredEmailChannelProvider',
         'options': {
             # list out all link resolvers
@@ -203,5 +220,5 @@ NOTIFICATION_DIGEST_EMAIL_CSS = 'css/email_digests.css'
 
 try:
     from local_settings import *
-except ImportError:
+except Exception:
     pass
