+++ testserver/templates/index.html
@@ -33,6 +33,11 @@
                 <option value='{{t.name}}'>{{t.name}}</option>
             {% endfor %}
         </select>
+        <label>Dispatch Channel</label><select name='notification_channel' >
+            <option value=''>----default----</option>
+            <option value='durable'>durable</option>
+            <option value='triggered-email'>triggered-email</option>
+        </select>
         <input type='submit' name='add_notifications' value='add a notification' />
         <br /><br />
         <input type='submit' name='send_digest' value='Send an unread Notifications Daily Digest' />&nbsp;NOTE: digest may or may not be sent depending on the preference setting of logged in user
